+++ src/lib/custom-providers.tsx
@@ -1,5 +1,9 @@
 'use client';
 
+import { useEffect, useState } from 'react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ThemeProvider as NextThemesProvider } from 'next-themes';
+import { Toaster } from 'react-hot-toast';
 import CreateChannelModal from '@/components/modal/CreateChannelModal';
 import CreateServerModal from '@/components/modal/CreateServerModal';
 import { DeleteChannelModal } from '@/components/modal/DeleteChannelModal';
@@ -10,7 +14,6 @@ import InviteServerModal from '@/components/modal/InviteServerModal';
 import { LeaveServerModal } from '@/components/modal/LeaveServerModal';
 import MembersModal from '@/components/modal/MembersModal';
 import { MessageFileModal } from '@/components/modal/MessageFileModal';
-import { useEffect, useState } from 'react';
 
 export const ModalProvider = () => {
   const [isMounted, setIsMounted] = useState(false);
@@ -38,3 +41,17 @@ export const ModalProvider = () => {
     </>
   );
 };
+
+export default function CustomProviders({ children }: { children: React.ReactNode }) {
+  const [queryClient] = useState(() => new QueryClient());
+
+  return (
+    <QueryClientProvider client={queryClient}>
+      <NextThemesProvider attribute="class" defaultTheme="system" enableSystem={false} storageKey="game-cord-theme">
+        <ModalProvider />
+        <Toaster />
+        {children}
+      </NextThemesProvider>
+    </QueryClientProvider>
+  );
+}
