--- src/providers/CustomReactQueryProvider.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { useState } from 'react';
-
-const CustomReactQueryProvider = ({ children }: { children: React.ReactNode }) => {
-  const [queryClient] = useState(() => new QueryClient());
-
-  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
-};
-
-export default CustomReactQueryProvider;
