--- src/providers/ToastProvider.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import { Toaster } from 'react-hot-toast';
-
-const ToastProvider = () => {
-  return <Toaster />;
-};
-
-export default ToastProvider;
